requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
matplotlib
scikit-learn
joblib
apache-airflow==3.0.0
